Enable this service to run on a different host to the proxy server
In order for Gunicorn to recognise the X-Forwarded-Proto header set by an Nginx reverse proxy (which tells it whether the originating request was on http or https), we need to tell it what IPs are allowed to set that header (it only allows localhost by default). This is important for when we need to generate absolute URLs - e.g. for redirects from Hydra during login. Without this capability, Flask will think the absolute URL is http which will not match the https redirect URLs configured for OAuth2 clients in Hydra.